refactor code and add __main__ module
