+++ docargs/__main__.py
@@ -1,0 +1,4 @@
+"""Module allowing for ``python -m docargs ...``."""
+from .cli import cli
+
+cli()
